pyproject.toml
[tool.poetry]
name = "django-boilerplate"
version = "0.1.0"
description = ""
authors = ["Abdullah zulfiqar <abdullahzulfiqar653@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "django_boilerplate"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0"
celery = "^5.3.6"
django-split-settings = "^1.2.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pre-commit = "^3.6.0"


[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
